// src/models/observation-model.ts
export type Attachment = {
  AttachmentId?: number;
  GeoHazardName?: string; //GeoHazard name
  RegistrationName?: string; //Registration name
  //UrlFormats	{...}
  Url?: string; //Full original image url
  Photographer?: string; //maxLength: 60 Navn på fotograf.
  Copyright?: string; //maxLength: 60 Rettigheter til bilde.
  Aspect?: number; //Hvilken himmelretning peker bilde. Gis i grader slik gitt på kompass. 0 er nord og 90 er øst osv.
  GeoHazardTID?: number; //Sett naturfare. Tabellen brukes av alle naturfarer (snø, jord, vann, is). The GeoHazardKD unique identifier
  RegistrationTID?: number; //Hva er bildet av. Dette feltet relaterer bildet til en observasjonstype. Feks værobservasjon, faretegn, osv. The RegistrationKD unique identifier
  Comment?: string; //maxLength: 2048 Kommentarfelt for bildet. F.eks for å beskrive det.
  AttachmentMimeType?: string;
  IsMainAttachment?: boolean; //Om bildet skal vises først i registreringen, eller ikke
  CropImage?: boolean;
};
// ...
export type Url = {
  UrlDescription?: string; //maxLength: 1024 UrlDescription
  UrlLine?: string; //maxLength: 1024 UrlLine
};